# docs/openapi.yaml
openapi: 3.1.0
info:
  title: XLS-66 Lending Protocol Reference App
  description: |
    HTTP API for the XRPL Lending Protocol (XLS-66) + Single Asset Vault
    (XLS-65) reference implementation. Every route that mutates on-chain
    state maps 1:1 to an XRPL transaction and re-queries the ledger before
    returning.

    ## Conventions

    - **Auth** — handled by Auth0 Universal Login. Clients obtain a session
      by visiting `/auth/login` (browser only) — there is no programmatic
      login endpoint. The session is stored in an httpOnly cookie `xls66-auth`.
      Every `/api/*` route is gated by `src/middleware.ts` and returns
      `401 Unauthorized` without a valid cookie.
    - **Amounts** — drops (integer string) for XRP; decimal string for IOU
      and MPT. See the README for the scaling conventions.
    - **Rates** — UI-facing rates are % or basis points; on-ledger they are
      1/10 bps. The API accepts the UI form and scales at the boundary.
    - **Content type** — `application/json` for all request/response bodies.
  version: 1.0.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: XLS-66 reference app
servers:
  - url: http://localhost:3000
    description: Local dev server
...
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: xls66-auth
      description: httpOnly cookie set by `POST /api/session`. Value is the Session document's MongoDB ObjectId.
  parameters:
...
    VaultIdPath:
      name: id
      in: path
      required: true
      schema: { type: string, description: "Hash256 of the Vault ledger entry" }
  schemas:
    Role:
      type: string
      enum: [broker, depositor, borrower, issuer]
...
    Vault:
      description: MongoDB cache row for a vault. Combined with `onLedger` for a full view.
      type: object
      required: [_id, sessionId, vaultId, ownerAddress]
      properties:
        _id: { type: string }
        sessionId: { type: string }
        vaultId: { type: string }
        ownerAddress: { type: string }
        asset:
          type: object
          properties:
            currency: { type: string }
            issuer: { type: string }
            mptIssuanceId: { type: string }
        totalDeposited: { type: string }
        sharesMinted: { type: string }
        status: { type: string, enum: [active, deleted] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        onLedger:
          description: Live `vault_info` response, amount fields unscaled for MPT.
          type: object
          additionalProperties: true
    Loan:
      description: MongoDB cache row for a loan. Cached terms + latest synced state.
      type: object
      required: [_id, sessionId, loanId, loanBrokerId, borrowerAddress]
      properties:
        _id: { type: string }
        sessionId: { type: string }
        loanId: { type: string }
        loanBrokerId: { type: string }
        borrowerAddress: { type: string }
        principalRequested: { type: string }
        interestRate: { type: integer, description: "Stored as basis points." }
        paymentTotal: { type: integer }
        paymentInterval: { type: integer, description: "Seconds" }
        gracePeriod: { type: integer, description: "Seconds" }
        originationFee: { type: string }
        serviceFee: { type: string }
        paymentsRemaining: { type: integer }
        principalOutstanding: { type: string }
        status: { type: string, enum: [active, repaid, defaulted, closed] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    TxResult:
      description: Pass-through of the rippled `tx` submission result. Only the hash is canonical; everything else is for debugging.
      type: object
      additionalProperties: true
      properties:
        hash: { type: string }
        ledger_index: { type: integer }
        meta:
          type: object
          additionalProperties: true
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
  responses:
    Unauthorized:
      description: Missing or invalid `xls66-auth` cookie.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Referenced entity not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Validation error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ServerError:
      description: Unhandled error or rippled failure.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
paths:
...
  /api/session/transfer:
    post:
      tags: [Session]
      summary: Peer transfer between role wallets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [from, to, amount]
              properties:
                from: { $ref: "#/components/schemas/Role" }
                to: { $ref: "#/components/schemas/Role" }
                amount: { type: string, description: "Human decimal. Amounts for XRP are interpreted as XRP (not drops)." }
                asset:
                  type: string
                  enum: [XRP, TUSD]
                  default: XRP
      responses:
        "200":
          description: Payment submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  result: { $ref: "#/components/schemas/TxResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/vault/{id}:
    get:
      tags: [Vault]
      summary: Vault detail
      parameters:
        - $ref: "#/components/parameters/VaultIdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  vault: { $ref: "#/components/schemas/Vault" }
                  onLedger:
                    description: "`vault_info` response with amount fields unscaled for MPT."
                    type: object
                    additionalProperties: true
        "500": { $ref: "#/components/responses/ServerError" }
...
  /api/loan:
    post:
      tags: [Loan]
      summary: Issue loan (LoanSet, multi-signed)
      description: |
        Broker signs first, borrower counter-signs via
        `xrpl.signLoanSetByCounterparty`. Amount fields (`principalRequested`,
        `originationFee`, `serviceFee`, `latePaymentFee`, `closePaymentFee`)
        are human values (drops for XRP, decimal for IOU/MPT); the server
        scales MPT values by `AssetScale` before building the tx.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                principalRequested: { type: string }
                interestRate: { type: integer, description: "Basis points (server × 10 for 1/10 bps on-chain)." }
                paymentTotal: { type: integer, minimum: 1, maximum: 120 }
                paymentInterval: { type: integer, description: "Seconds between payments (min 60)." }
                gracePeriod: { type: integer, description: "Seconds" }
                originationFee: { type: string }
                serviceFee: { type: string }
                latePaymentFee: { type: string }
                closePaymentFee: { type: string }
                overpaymentFee: { type: integer, description: "1/10 bps" }
                lateInterestRate: { type: integer, description: "1/10 bps" }
                closeInterestRate: { type: integer, description: "1/10 bps" }
                overpaymentInterestRate: { type: integer, description: "1/10 bps" }
                loanName: { type: string, maxLength: 64 }
                allowOverpayment:
                  type: boolean
                  description: "Sets tfLoanOverpayment on LoanSet (enables later overpayment)."
      responses:
        "201":
          description: Loan created
          content:
            application/json:
              schema:
                type: object
                properties:
                  loanId: { type: string }
                  result: { $ref: "#/components/schemas/TxResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500":
          description: Loan submission failed (e.g. `tecPRECISION_LOSS`, `tecLIMIT_EXCEEDED`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      tags: [Loan]
      summary: List loans
      description: Returns all session loans with on-chain sync. Updates DB `paymentsRemaining` / `principalOutstanding` / `status` from `ledger_entry`; unscales MPT amounts.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  loans:
                    type: array
                    items: { $ref: "#/components/schemas/Loan" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...